+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Present aliases for the RustSec ID and related advisories in the overview table (#1).
 
+### Changed
+
+* Setting `denyWarnings` will now pass `--deny warnings` to cargo audit.
+
 ## [1.0.1] - 2022-08-09
 
 ### Added